requirements.txt
tensorflow 
pandas 
dvc 
mlflow
notebook
numpy
matplotlib
seaborn 
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
scipy
Flask
Flask-Cors
gdown
-e .